Unbreak loading profile images, mostly
Now that TL uses CSS classes for profile icons, it's slightly trickier
to get the proper icon image. Attempted a fix which is all kinds of ugly
(it just guesses class.gif as the image name). Also used some Java XPath
stuff instead of HTMLcleaner which doesn't support the XPath contains
function, probably in an inefficient manner. But now it works again.

(icons are tiny for some reason, will need to fix)